Adding in the rng for the energy, adjusting histograms
run.bash:
Setting up for a mass run over several production files to update with the
new histograms.

Calibrator.cpp:
Updated the CsI calibration with some new values.

ProjectNSub.C:
Adjusting the ranges of the histograms to match with the originals.

he6scan.cpp:
Adding in the random number for the energy. Necessary to remove artifacts
when applying the calibrations. Adding in a calibrated energy histogram.
Removing some unnecessary histograms. Adjusting some of the cuts on when
to plot things. Changing where the time is updated, before setting new
time instead of after (shouldn't change things).

Signed-off-by: S.V. Paulauskas <stanpaulauskas@gmail.com>